major changes and fixed several bugs
1. change float32 to float64.

2. change all randomly initialized weight and bias into float64(0.1). (Original code is commented at corresponding position)

3.  Linear forward input initialization

4. Convolution layer backward function changed: do not initialize both the KGradient and BGradient fields to be all zeros. Keep them as its original value.

5. convo backward. Divide KGradient by bxfloat64 only once at the end instead of each time.

+++ main.go
@@ -9,7 +9,7 @@ func main() {
 	//training process
 	fmt.Println("Start training process")
 
-	learning_rate := float32(0.01)
+	learning_rate := float64(0.01)
 	batch_size := 3
 	num_epoch := 3
 	path := "MINST"
@@ -28,31 +28,30 @@ func main() {
 	
 	/*
 	fmt.Println("forward test starts here")
-
 	// define a new tensor to feed our CNN
-	input := make([][][][]float32, 3) // number of images, number of inChannel, height, width
+	input := make([][][][]float64, 3) // number of images, number of inChannel, height, width
 	for b := 0; b < 3; b++ {
-		input[b] = make([][][]float32, 1)
-		input[b][0] = make([][]float32, 5)
+		input[b] = make([][][]float64, 1)
+		input[b][0] = make([][]float64, 5)
 	}
 	// first image
-	input[0][0][0] = []float32{1,2,3,4,5}
-	input[0][0][1] = []float32{2,3,4,5,6}
-	input[0][0][2] = []float32{3,4,5,6,7}
-	input[0][0][3] = []float32{4,5,6,7,8}
-	input[0][0][4] = []float32{5,6,7,8,9}
+	input[0][0][0] = []float64{1,2,3,4,5}
+	input[0][0][1] = []float64{2,3,4,5,6}
+	input[0][0][2] = []float64{3,4,5,6,7}
+	input[0][0][3] = []float64{4,5,6,7,8}
+	input[0][0][4] = []float64{5,6,7,8,9}
 	// second image
-	input[1][0][0] = []float32{1,2,3,4,5}
-	input[1][0][1] = []float32{2,3,4,5,6}
-	input[1][0][2] = []float32{3,4,5,6,7}
-	input[1][0][3] = []float32{4,5,6,7,8}
-	input[1][0][4] = []float32{5,6,7,8,9}
+	input[1][0][0] = []float64{1,2,3,4,5}
+	input[1][0][1] = []float64{2,3,4,5,6}
+	input[1][0][2] = []float64{3,4,5,6,7}
+	input[1][0][3] = []float64{4,5,6,7,8}
+	input[1][0][4] = []float64{5,6,7,8,9}
 	// third image
-	input[2][0][0] = []float32{1,2,3,4,5}
-	input[2][0][1] = []float32{2,3,4,5,6}
-	input[2][0][2] = []float32{3,4,5,6,7}
-	input[2][0][3] = []float32{4,5,6,7,8}
-	input[2][0][4] = []float32{5,6,7,8,9}
+	input[2][0][0] = []float64{1,2,3,4,5}
+	input[2][0][1] = []float64{2,3,4,5,6}
+	input[2][0][2] = []float64{3,4,5,6,7}
+	input[2][0][3] = []float64{4,5,6,7,8}
+	input[2][0][4] = []float64{5,6,7,8,9}
 	fmt.Println("the input sample of three images as a batch:")
 	fmt.Println(input[0])
 	fmt.Println(input[1])
@@ -71,22 +70,22 @@ func main() {
 	fmt.Println("")
 
 	// initialize the kernel for a convolution layer
-	kernelTest := make([][][][]float32, 2) // height, width, # inChanel = 1, # kernels(OutChannel)
+	kernelTest := make([][][][]float64, 2) // height, width, # inChanel = 1, # kernels(OutChannel)
 	for i := 0; i < 2; i++ {
-		kernelTest[i] = make([][][]float32, 2)
+		kernelTest[i] = make([][][]float64, 2)
 		for j := 0; j < 2; j++ {
-			kernelTest[i][j] = make([][]float32, 1)
-			kernelTest[i][j][0] = make([]float32, 4)
+			kernelTest[i][j] = make([][]float64, 1)
+			kernelTest[i][j][0] = make([]float64, 4)
 		}
 	}
-	kernelTest[0][0][0] = []float32{1,1,1,1}
-	kernelTest[0][1][0] = []float32{2,2,2,2}
-	kernelTest[1][0][0] = []float32{2,2,2,2}
-	kernelTest[1][1][0] = []float32{3,3,3,3}
+	kernelTest[0][0][0] = []float64{1,1,1,1}
+	kernelTest[0][1][0] = []float64{2,2,2,2}
+	kernelTest[1][0][0] = []float64{2,2,2,2}
+	kernelTest[1][1][0] = []float64{3,3,3,3}
 	fmt.Println("convolution layer's kernel is:")
 	fmt.Println(kernelTest)
 	conv_1 := InitializeConvAssigned(kernelTest, 0, 1, 3)
-	conv_1.Bias = []float32{1,2,3,4}
+	conv_1.Bias = []float64{1,2,3,4}
 	fmt.Println("")
 
 	// forward convolution, manually checked for correction
@@ -110,34 +109,34 @@ func main() {
 	fmt.Println("")
 
 	// backward convolution
-	deltaC := make([][][][]float32, 3)
+	deltaC := make([][][][]float64, 3)
 
 	for b := 0; b < 3; b++ {
-		deltaC[b] = make([][][]float32, 4)
+		deltaC[b] = make([][][]float64, 4)
 
-		deltaC[b][0] = make([][]float32, 4)
-		deltaC[b][0][0] = []float32{1,2,3,4}
-		deltaC[b][0][1] = []float32{2,3,4,5}
-		deltaC[b][0][2] = []float32{3,4,5,6}
-		deltaC[b][0][3] = []float32{4,5,6,7}
+		deltaC[b][0] = make([][]float64, 4)
+		deltaC[b][0][0] = []float64{1,2,3,4}
+		deltaC[b][0][1] = []float64{2,3,4,5}
+		deltaC[b][0][2] = []float64{3,4,5,6}
+		deltaC[b][0][3] = []float64{4,5,6,7}
 
-		deltaC[b][1] = make([][]float32, 4)
-		deltaC[b][1][0] = []float32{3,4,5,6}
-		deltaC[b][1][1] = []float32{4,5,6,7}
-		deltaC[b][1][2] = []float32{5,6,7,8}
-		deltaC[b][1][3] = []float32{6,7,8,9}
+		deltaC[b][1] = make([][]float64, 4)
+		deltaC[b][1][0] = []float64{3,4,5,6}
+		deltaC[b][1][1] = []float64{4,5,6,7}
+		deltaC[b][1][2] = []float64{5,6,7,8}
+		deltaC[b][1][3] = []float64{6,7,8,9}
 
-		deltaC[b][2] = make([][]float32, 4)
-		deltaC[b][2][0] = []float32{5,6,7,8}
-		deltaC[b][2][1] = []float32{6,7,8,9}
-		deltaC[b][2][2] = []float32{7,8,9,10}
-		deltaC[b][2][3] = []float32{8,9,10,11}
+		deltaC[b][2] = make([][]float64, 4)
+		deltaC[b][2][0] = []float64{5,6,7,8}
+		deltaC[b][2][1] = []float64{6,7,8,9}
+		deltaC[b][2][2] = []float64{7,8,9,10}
+		deltaC[b][2][3] = []float64{8,9,10,11}
 
-		deltaC[b][3] = make([][]float32, 4)
-		deltaC[b][3][0] = []float32{7,8,9,10}
-		deltaC[b][3][1] = []float32{8,9,10,11}
-		deltaC[b][3][2] = []float32{9,10,11,12}
-		deltaC[b][3][3] = []float32{10,11,12,13}
+		deltaC[b][3] = make([][]float64, 4)
+		deltaC[b][3][0] = []float64{7,8,9,10}
+		deltaC[b][3][1] = []float64{8,9,10,11}
+		deltaC[b][3][2] = []float64{9,10,11,12}
+		deltaC[b][3][3] = []float64{10,11,12,13}
 	}
 	convBack := conv_1.Backward(deltaC, 0.01)
 	fmt.Println("after convolution backward, delta is:")
@@ -193,7 +192,7 @@ func main() {
 	fmt.Println("")
 
 	// newLinear: fully-connected layer
-	linear :=  NewLinearAssigned(16, 10) // all weight and bias equal float32(1)
+	linear :=  NewLinearAssigned(16, 10) // all weight and bias equal float64(1)
 	pool_1_output_reshaped := Reshape4Dto2D(pool_1_output) // has shape: batchSize by everythingElseFlattened
 	linear_output := linear.Forward(pool_1_output_reshaped) // shape: 3 by 10
 	fmt.Println("output from fully connected layer of size 3*10:")
@@ -206,10 +205,10 @@ func main() {
 	fmt.Println("")
 
 	// sofmax layer
-	batchLabel := make([][]float32, 3)
-	batchLabel[0] = []float32{1,0,0,0,0,0,0,0,0,0}
-	batchLabel[1] = []float32{0,1,0,0,0,0,0,0,0,0}
-	batchLabel[2] = []float32{0,0,1,0,0,0,0,0,0,0}
+	batchLabel := make([][]float64, 3)
+	batchLabel[0] = []float64{1,0,0,0,0,0,0,0,0,0}
+	batchLabel[1] = []float64{0,1,0,0,0,0,0,0,0,0}
+	batchLabel[2] = []float64{0,0,1,0,0,0,0,0,0,0}
 	loss, delta := SoftmaxCalLoss(linear_output, batchLabel)
 	fmt.Println("output from softmax layer:")
 	fmt.Println("loss:", loss)
@@ -221,18 +220,18 @@ func main() {
 
 	// linear backward
 	linearB := NewLinearAssigned(4, 3)
-	linearB.x = make([][]float32, 2)
-	linearB.x[0] = []float32{1,2,3,4}
-	linearB.x[1] = []float32{2,3,4,5}
-	linearB.W[0] = []float32{1,2,3}
-	linearB.W[1] = []float32{2,3,4}
-	linearB.W[2] = []float32{3,4,5}
-	linearB.W[3] = []float32{4,5,6}
-	linearB.b = []float32{1,2,3}
-	deltab := make([][]float32, 2)
-	deltab[0] = []float32{1,2,3}
-	deltab[1] = []float32{2,3,4}
-	deltabPrime := linearB.Backward(deltab, float32(0.01))
+	linearB.x = make([][]float64, 2)
+	linearB.x[0] = []float64{1,2,3,4}
+	linearB.x[1] = []float64{2,3,4,5}
+	linearB.W[0] = []float64{1,2,3}
+	linearB.W[1] = []float64{2,3,4}
+	linearB.W[2] = []float64{3,4,5}
+	linearB.W[3] = []float64{4,5,6}
+	linearB.b = []float64{1,2,3}
+	deltab := make([][]float64, 2)
+	deltab[0] = []float64{1,2,3}
+	deltab[1] = []float64{2,3,4}
+	deltabPrime := linearB.Backward(deltab, float64(0.01))
 	fmt.Println("after linear backward:")
 	fmt.Println(deltabPrime[0])
 	fmt.Println(deltabPrime[1])
@@ -248,7 +247,7 @@ func main() {
 }
 
 /*
-func InitializeConvAssigned(kernel [][][][]float32, pad, stride, numImages int) *Convolution {
+func InitializeConvAssigned(kernel [][][][]float64, pad, stride, numImages int) *Convolution {
 
 	var convLayer Convolution // one convolution layer
 
@@ -267,30 +266,30 @@ func InitializeConvAssigned(kernel [][][][]float32, pad, stride, numImages int) 
 	convLayer.Kernel = kernel
 
 	// initialize "Bias"
-	convLayer.Bias = make([]float32, kernelShape[3])
+	convLayer.Bias = make([]float64, kernelShape[3])
 	for i := 0; i < len(convLayer.Bias); i++ {
-		convLayer.Bias[i] = float32(0.1) * float32(i)
+		convLayer.Bias[i] = float64(0.1) * float64(i)
 	}
 	fmt.Println("convolution layer's bias is:")
 	fmt.Println(convLayer.Bias)
 
 	// initialize "KGradient"
-	convLayer.KGradient = make([][][][]float32, kernelShape[0])
+	convLayer.KGradient = make([][][][]float64, kernelShape[0])
 	for i := 0; i < kernelShape[0]; i++ {
-		convLayer.KGradient[i] = make([][][]float32, kernelShape[1])
+		convLayer.KGradient[i] = make([][][]float64, kernelShape[1])
 		for ii := 0; ii < kernelShape[1]; ii++ {
-			convLayer.KGradient[i][ii] = make([][]float32, kernelShape[2])
+			convLayer.KGradient[i][ii] = make([][]float64, kernelShape[2])
 			for iii := 0; iii < kernelShape[2]; iii++ {
-				convLayer.KGradient[i][ii][iii] = make([]float32, kernelShape[3])
+				convLayer.KGradient[i][ii][iii] = make([]float64, kernelShape[3])
 			}
 		}
 	}
 
 	// initialize "BGradient"
-	convLayer.BGradient = make([]float32, kernelShape[3])
+	convLayer.BGradient = make([]float64, kernelShape[3])
 
 	// initialize "ImageCol"
-	convLayer.ImageCol = make([][][][]float32, numImages)
+	convLayer.ImageCol = make([][][][]float64, numImages)
 
 	pointer := &convLayer
 	return pointer
@@ -300,20 +299,20 @@ func InitializeConvAssigned(kernel [][][][]float32, pad, stride, numImages int) 
 func NewLinearAssigned(inChannel, outChannel int) *Linear {
 
 	// Initialize the weights and biases with random values (gradients will be zero)
-	W := make([][]float32, inChannel)
-	WGradient := make([][]float32, inChannel)
+	W := make([][]float64, inChannel)
+	WGradient := make([][]float64, inChannel)
 	for i := range W {
-		W[i] = make([]float32, outChannel)
-		WGradient[i] = make([]float32, outChannel)
+		W[i] = make([]float64, outChannel)
+		WGradient[i] = make([]float64, outChannel)
 		for j := range W[i] {
-			W[i][j] = float32(1)
+			W[i][j] = float64(1)
 		}
 	}
 
-	b := make([]float32, outChannel)
-	bGradient := make([]float32, outChannel)
+	b := make([]float64, outChannel)
+	bGradient := make([]float64, outChannel)
 	for i := range b {
-		b[i] = float32(1)
+		b[i] = float64(1)
 		bGradient[i] = 0
 	}
